+++ src/main/java/com/sopt/push/service/DeviceTokenService.java
@@ -5,8 +5,12 @@ import static com.sopt.push.common.Constants.TOKEN_PREFIX;
 import static com.sopt.push.common.Constants.USER_PREFIX;
 
 import com.sopt.push.domain.DeviceTokenEntity;
+import com.sopt.push.dto.UserTokenInfoDto;
+import com.sopt.push.enums.Platform;
 import com.sopt.push.repository.DeviceTokenRepository;
 import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
 
 public class DeviceTokenService {
 
@@ -43,4 +47,30 @@ public class DeviceTokenService {
 
     deviceTokenRepository.delete(tokenPk, userSk);
   }
+
+    public List<DeviceTokenEntity> findUserByTokenIds(List<String> deviceTokens) {
+        List<DeviceTokenEntity> result = new ArrayList<>();
+        for (String deviceToken : deviceTokens) {
+            deviceTokenRepository.findByDeviceToken(deviceToken).ifPresent(result::add);
+        }
+        return result;
+    }
+
+    public UserTokenInfoDto mapDeviceTokenEntityToInfoDto(DeviceTokenEntity deviceTokenEntity) {
+        String deviceToken =
+                deviceTokenEntity.getPk().startsWith(TOKEN_PREFIX)
+                        ? deviceTokenEntity.getPk().substring(TOKEN_PREFIX.length())
+                        : deviceTokenEntity.getPk();
+        String userId =
+                deviceTokenEntity.getSk().startsWith(USER_PREFIX)
+                        ? deviceTokenEntity.getSk().substring(USER_PREFIX.length())
+                        : deviceTokenEntity.getSk();
+
+        return new UserTokenInfoDto(
+                userId,
+                deviceToken,
+                deviceTokenEntity.getEndpointArn(),
+                Platform.fromValue(deviceTokenEntity.getPlatform()),
+                deviceTokenEntity.getSubscriptionArn());
+    }
 }
